% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ch19to38}
\alias{ch19to38}
\title{liftOver support for erma data in hg19}
\format{chain imported using rtracklayer}
\usage{
ch19to38
}
\description{
liftOver support for erma data in hg19
}
\keyword{datasets}
