% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lgenGWC_17}
\alias{lgenGWC_17}
\title{illustrative excerpt from EBI GWAS catalog for jan 2019; limit to chr17}
\format{gwaswloc instance as defined in gwascat package}
\usage{
lgenGWC_17
}
\description{
illustrative excerpt from EBI GWAS catalog for jan 2019; limit to chr17
}
\examples{
lungGen::lgenGWC_17
}
\keyword{datasets}
