% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenoToHits.R
\name{phenoToHits}
\alias{phenoToHits}
\title{use MAPPED_TRAIT_URI to find hits after mapping text to EBI ontology tag}
\usage{
phenoToHits(interm, gwc)
}
\arguments{
\item{interm}{character(1) term of interest in English}

\item{gwc}{instance of gwaswloc}
}
\description{
use MAPPED_TRAIT_URI to find hits after mapping text to EBI ontology tag
}
\examples{
data(lgenGWC_17)
ph = phenoToHits("pulmonary", lgenGWC_17)
ph
table(ph$MAPPED_TRAIT)
}
