% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{predGR}
\alias{predGR}
\title{location, rsid and predicted probability of phenorelevance based on jan 30 2019 model}
\format{GRanges instance, GRCh38 addresses}
\usage{
predGR
}
\description{
location, rsid and predicted probability of phenorelevance based on jan 30 2019 model
}
\examples{
head(predGR)
}
\keyword{datasets}
