% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{taggedPhenoDF}
\alias{taggedPhenoDF}
\title{mapping from MAPPED_TRAIT_URI tags to natural language terms}
\format{data.frame}
\usage{
taggedPhenoDF
}
\description{
mapping from MAPPED_TRAIT_URI tags to natural language terms
}
\examples{
lungGen::taggedPhenoDF
}
\keyword{datasets}
